remove all context prop
