+++ src/main/java/com/uber/cadence/internal/context/ContextThreadLocal.java
@@ -24,13 +24,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-/** This class holds the current set of context propagators. */
+/** This class holds the current set of context propagators */
 public class ContextThreadLocal {
-
-  private static final Logger log = LoggerFactory.getLogger(ContextThreadLocal.class);
 
   private static WorkflowThreadLocal<List<ContextPropagator>> contextPropagators =
       WorkflowThreadLocal.withInitial(
@@ -41,7 +37,7 @@ public class ContextThreadLocal {
             }
           });
 
-  /** Sets the list of context propagators for the thread. */
+  /** Sets the list of context propagators for the thread */
   public static void setContextPropagators(List<ContextPropagator> propagators) {
     if (propagators == null || propagators.isEmpty()) {
       return;
@@ -61,11 +57,6 @@ public class ContextThreadLocal {
     return contextData;
   }
 
-  /**
-   * Injects the context data into the thread for each configured context propagator.
-   *
-   * @param contextData The context data received from the server.
-   */
   public static void propagateContextToCurrentThread(Map<String, Object> contextData) {
     if (contextData == null || contextData.isEmpty()) {
       return;
@@ -73,46 +64,6 @@ public class ContextThreadLocal {
     for (ContextPropagator propagator : contextPropagators.get()) {
       if (contextData.containsKey(propagator.getName())) {
         propagator.setCurrentContext(contextData.get(propagator.getName()));
-      }
-    }
-  }
-
-  /** Calls {@link ContextPropagator#setUp()} for each propagator. */
-  public static void setUpContextPropagators() {
-    for (ContextPropagator propagator : contextPropagators.get()) {
-      try {
-        propagator.setUp();
-      } catch (Throwable t) {
-        // Don't let an error in one propagator block the others
-        log.error("Error calling setUp() on a contextpropagator", t);
-      }
-    }
-  }
-
-  /**
-   * Calls {@link ContextPropagator#onError(Throwable)} for each propagator.
-   *
-   * @param t The Throwable that caused the workflow/activity to finish.
-   */
-  public static void onErrorContextPropagators(Throwable t) {
-    for (ContextPropagator propagator : contextPropagators.get()) {
-      try {
-        propagator.onError(t);
-      } catch (Throwable t1) {
-        // Don't let an error in one propagator block the others
-        log.error("Error calling onError() on a contextpropagator", t1);
-      }
-    }
-  }
-
-  /** Calls {@link ContextPropagator#finish()} for each propagator. */
-  public static void finishContextPropagators() {
-    for (ContextPropagator propagator : contextPropagators.get()) {
-      try {
-        propagator.finish();
-      } catch (Throwable t) {
-        // Don't let an error in one propagator block the others
-        log.error("Error calling finish() on a contextpropagator", t);
       }
     }
   }
