+++ src/main/java/com/uber/cadence/internal/worker/ActivityWorker.java
@@ -41,7 +41,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.CancellationException;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
 import org.apache.thrift.TException;
 import org.slf4j.MDC;
 
@@ -162,11 +161,7 @@ public class ActivityWorker extends SuspendableWorkerBase {
         Stopwatch sw = metricsScope.timer(MetricsType.ACTIVITY_RESP_LATENCY).start();
         sendReply(task, new Result(null, null, cancelledRequest), metricsScope);
         sw.stop();
-        onErrorContextPropagation(e);
-      } catch (Exception e) {
-        onErrorContextPropagation(e);
       } finally {
-        finishContextPropagation();
         MDC.remove(LoggerTag.ACTIVITY_ID);
         MDC.remove(LoggerTag.ACTIVITY_TYPE);
         MDC.remove(LoggerTag.WORKFLOW_ID);
@@ -193,31 +188,7 @@ public class ActivityWorker extends SuspendableWorkerBase {
               });
 
       for (ContextPropagator propagator : options.getContextPropagators()) {
-        // Only send the context propagator the fields that belong to them
-        // Change the map from MyPropagator:foo -> bar to foo -> bar
-        Map<String, byte[]> filteredData =
-            headerData
-                .entrySet()
-                .stream()
-                .filter(e -> e.getKey().startsWith(propagator.getName()))
-                .collect(
-                    Collectors.toMap(
-                        e -> e.getKey().substring(propagator.getName().length() + 1),
-                        Map.Entry::getValue));
-        propagator.setCurrentContext(propagator.deserializeContext(filteredData));
-        propagator.setUp();
-      }
-    }
-
-    void onErrorContextPropagation(Exception error) {
-      for (ContextPropagator propagator : options.getContextPropagators()) {
-        propagator.onError(error);
-      }
-    }
-
-    void finishContextPropagation() {
-      for (ContextPropagator propagator : options.getContextPropagators()) {
-        propagator.finish();
+        propagator.setCurrentContext(propagator.deserializeContext(headerData));
       }
     }
